TINY-9635: Refactor sanitization code into Sanitization module
