+++ modules/tinymce/src/core/main/ts/api/html/DomParser.ts
@@ -1,6 +1,4 @@
-import { Arr, Fun, Obj, Strings, Type } from '@ephox/katamari';
-import { Attribute, NodeTypes, Remove, Replication, SugarElement } from '@ephox/sugar';
-import createDompurify, { Config, DOMPurifyI } from 'dompurify';
+import { Arr, Fun, Obj, Type } from '@ephox/katamari';
 
 import * as TransparentElements from '../../content/TransparentElements';
 import * as NodeType from '../../dom/NodeType';
@@ -12,8 +10,8 @@ import * as ParserFilters from '../../html/ParserFilters';
 import { isEmpty, isLineBreakNode, isPaddedWithNbsp, paddEmptyNode } from '../../html/ParserUtils';
 import { BlobCache } from '../file/BlobCache';
 import Tools from '../util/Tools';
-import * as URI from '../util/URI';
 import AstNode from './Node';
+import { getSanitizer } from './Sanitization';
 import Schema, { getTextRootBlockElements, SchemaMap, SchemaRegExpMap } from './Schema';
 
 /**
@@ -83,179 +81,6 @@ interface DomParser {
 }
 
 type WalkerCallback = (node: AstNode) => void;
-
-const basePurifyConfig: Config = {
-  IN_PLACE: true,
-  ALLOW_UNKNOWN_PROTOCOLS: true,
-  // Deliberately ban all tags and attributes by default, and then un-ban them on demand in hooks
-  // #comment and #cdata-section are always allowed as they aren't controlled via the schema
-  // body is also allowed due to the DOMPurify checking the root node before sanitizing
-  ALLOWED_TAGS: [ '#comment', '#cdata-section', 'body' ],
-  ALLOWED_ATTR: []
-};
-
-// A list of attributes that should be filtered further based on the parser settings
-const filteredUrlAttrs = Tools.makeMap('src,href,data,background,action,formaction,poster,xlink:href');
-const internalElementAttr = 'data-mce-type';
-
-const getPurifyConfig = (settings: DomParserSettings, mimeType: string): Config => {
-  const config = { ...basePurifyConfig };
-
-  // Set the relevant parser mimetype
-  config.PARSER_MEDIA_TYPE = mimeType;
-
-  // Allow any URI when allowing script urls
-  if (settings.allow_script_urls) {
-    config.ALLOWED_URI_REGEXP = /.*/;
-  // Allow anything except javascript (or similar) URIs if all html data urls are allowed
-  } else if (settings.allow_html_data_urls) {
-    config.ALLOWED_URI_REGEXP = /^(?!(\w+script|mhtml):)/i;
-  }
-
-  return config;
-};
-
-const processNode = (node: Node, settings: DomParserSettings, schema: Schema, uid: number, evt?: createDompurify.SanitizeElementHookEvent): number => {
-  const validate = settings.validate;
-  const specialElements = schema.getSpecialElements();
-
-  // Pad conditional comments if they aren't allowed
-  if (node.nodeType === NodeTypes.COMMENT && !settings.allow_conditional_comments && /^\[if/i.test(node.nodeValue ?? '')) {
-    node.nodeValue = ' ' + node.nodeValue;
-  }
-
-  const lcTagName = evt?.tagName ?? node.nodeName.toLowerCase();
-
-  // Just leave non-elements such as text and comments up to dompurify
-  if (node.nodeType !== NodeTypes.ELEMENT || lcTagName === 'body') {
-    return uid;
-  }
-
-  // Construct the sugar element wrapper
-  const element = SugarElement.fromDom(node) as SugarElement<Element>;
-
-  // Determine if we're dealing with an internal attribute
-  const isInternalElement = Attribute.has(element, internalElementAttr);
-
-  // Cleanup bogus elements
-  const bogus = Attribute.get(element, 'data-mce-bogus');
-  if (!isInternalElement && Type.isString(bogus)) {
-    if (bogus === 'all') {
-      Remove.remove(element);
-    } else {
-      Remove.unwrap(element);
-    }
-    return uid;
-  }
-
-  // Determine if the schema allows the element and either add it or remove it
-  const rule = schema.getElementRule(lcTagName);
-  if (validate && !rule) {
-    // If a special element is invalid, then remove the entire element instead of unwrapping
-    if (Obj.has(specialElements, lcTagName)) {
-      Remove.remove(element);
-    } else {
-      Remove.unwrap(element);
-    }
-    return uid;
-  } else {
-    if (Type.isNonNullable(evt)) {
-      evt.allowedTags[lcTagName] = true;
-    }
-  }
-
-  // Validate the element using the attribute rules
-  if (validate && rule && !isInternalElement) {
-    // Fix the attributes for the element, unwrapping it if we have to
-    Arr.each(rule.attributesForced ?? [], (attr) => {
-      Attribute.set(element, attr.name, attr.value === '{$uid}' ? `mce_${uid++}` : attr.value);
-    });
-    Arr.each(rule.attributesDefault ?? [], (attr) => {
-      if (!Attribute.has(element, attr.name)) {
-        Attribute.set(element, attr.name, attr.value === '{$uid}' ? `mce_${uid++}` : attr.value);
-      }
-    });
-
-    // If none of the required attributes were found then remove
-    if (rule.attributesRequired && !Arr.exists(rule.attributesRequired, (attr) => Attribute.has(element, attr))) {
-      Remove.unwrap(element);
-      return uid;
-    }
-
-    // If there are no attributes then remove
-    if (rule.removeEmptyAttrs && Attribute.hasNone(element)) {
-      Remove.unwrap(element);
-      return uid;
-    }
-
-    // Change the node name if the schema says to
-    if (rule.outputName && rule.outputName !== lcTagName) {
-      Replication.mutate(element, rule.outputName as keyof HTMLElementTagNameMap);
-    }
-  }
-
-  return uid;
-};
-
-const shouldKeepAttribute = (settings: DomParserSettings, schema: Schema, tagName: string, attrName: string, attrValue: string): boolean =>
-  !(attrName in filteredUrlAttrs && URI.isInvalidUri(settings, attrValue, tagName)) &&
-  (!settings.validate || schema.isValid(tagName, attrName) || Strings.startsWith(attrName, 'data-') || Strings.startsWith(attrName, 'aria-'));
-
-const isRequiredAttributeOfInternalElement = (ele: Element, attrName: string): boolean =>
-  ele.hasAttribute(internalElementAttr) && (attrName === 'id' || attrName === 'class' || attrName === 'style');
-
-const isBooleanAttribute = (attrName: string, schema: Schema): boolean =>
-  attrName in schema.getBoolAttrs();
-
-const filterAttributes = (ele: Element, settings: DomParserSettings, schema: Schema): void => {
-  const { attributes } = ele;
-  for (let i = attributes.length - 1; i >= 0; i--) {
-    const attr = attributes[i];
-    const attrName = attr.name;
-    const attrValue = attr.value;
-    if (!shouldKeepAttribute(settings, schema, ele.tagName.toLowerCase(), attrName, attrValue) && !isRequiredAttributeOfInternalElement(ele, attrName)) {
-      ele.removeAttribute(attrName);
-    } else if (isBooleanAttribute(attrName, schema)) {
-      ele.setAttribute(attrName, attrName);
-    }
-  }
-};
-
-const setupPurify = (settings: DomParserSettings, schema: Schema): DOMPurifyI => {
-  const purify = createDompurify();
-  let uid = 0;
-
-  // We use this to add new tags to the allow-list as we parse, if we notice that a tag has been banned but it's still in the schema
-  purify.addHook('uponSanitizeElement', (ele, evt) => {
-    uid = processNode(ele, settings, schema, uid, evt);
-  });
-
-  // Let's do the same thing for attributes
-  purify.addHook('uponSanitizeAttribute', (ele, evt) => {
-    const tagName = ele.tagName.toLowerCase();
-    const { attrName, attrValue } = evt;
-
-    evt.keepAttr = shouldKeepAttribute(settings, schema, tagName, attrName, attrValue);
-
-    if (evt.keepAttr) {
-      evt.allowedAttributes[attrName] = true;
-
-      if (isBooleanAttribute(attrName, schema)) {
-        evt.attrValue = attrName;
-      }
-
-      // We need to tell DOMPurify to forcibly keep the attribute if it's an SVG data URI and svg data URIs are allowed
-      if (settings.allow_svg_data_urls && Strings.startsWith(attrValue, 'data:image/svg+xml')) {
-        evt.forceKeepAttr = true;
-      }
-    // For internal elements always keep the attribute if the attribute name is id, class or style
-    } else if (isRequiredAttributeOfInternalElement(ele, attrName)) {
-      evt.forceKeepAttr = true;
-    }
-  });
-
-  return purify;
-};
 
 const transferChildren = (parent: AstNode, nativeParent: Node, specialElements: SchemaRegExpMap) => {
   const parentName = parent.name;
@@ -444,7 +269,7 @@ const DomParser = (settings: DomParserSettings = {}, schema = Schema()): DomPars
   };
 
   const parser = new DOMParser();
-  const purify = setupPurify(defaultedSettings, schema);
+  const sanitize = getSanitizer(defaultedSettings, schema);
 
   const parseAndSanitizeWithContext = (html: string, rootName: string, format: string = 'html'): Element => {
     const mimeType = format === 'xhtml' ? 'application/xhtml+xml' : 'text/html';
@@ -455,24 +280,7 @@ const DomParser = (settings: DomParserSettings = {}, schema = Schema()): DomPars
     // If parsing XHTML then the content must contain the xmlns declaration, see https://www.w3.org/TR/xhtml1/normative.html#strict
     const wrappedHtml = format === 'xhtml' ? `<html xmlns="http://www.w3.org/1999/xhtml"><head></head><body>${content}</body></html>` : `<body>${content}</body>`;
     const body = parser.parseFromString(wrappedHtml, mimeType).body;
-
-    if (defaultedSettings.sanitize) {
-      // Sanitize the content
-      purify.sanitize(body, getPurifyConfig(defaultedSettings, mimeType));
-      purify.removed = [];
-    } else {
-      // eslint-disable-next-line no-bitwise
-      const nodeIterator = document.createNodeIterator(body, NodeFilter.SHOW_ELEMENT | NodeFilter.SHOW_COMMENT | NodeFilter.SHOW_TEXT);
-      let uid = 0;
-      let node;
-      while ((node = nodeIterator.nextNode())) {
-        uid = processNode(node, defaultedSettings, schema, uid);
-        if (NodeType.isElement(node)) {
-          filterAttributes(node, defaultedSettings, schema);
-        }
-      }
-    }
-
+    sanitize(body, mimeType);
     return isSpecialRoot ? body.firstChild as Element : body;
   };
 
@@ -549,7 +357,7 @@ const DomParser = (settings: DomParserSettings = {}, schema = Schema()): DomPars
   };
 
   const isWrappableNode = (blockElements: SchemaMap, node: AstNode) => {
-    const isInternalElement = Type.isString(node.attr(internalElementAttr));
+    const isInternalElement = Type.isString(node.attr('data-mce-type'));
     const isInlineElement = node.type === 1 && (!Obj.has(blockElements, node.name) && !TransparentElements.isTransparentAstBlock(schema, node));
 
     return node.type === 3 || (isInlineElement && !isInternalElement);
